完善了类型转换

+++ shop_assistant/src/c/C_goods.h
@@ -9,13 +9,13 @@ typedef struct {
 	Date C_producedDate;//生产日期
 	Date C_entryDate;//入库日期
 	Date C_rotDate;//过期日期
-	int C_in;//进价
+	float C_in;//进价
 	int C_shelfDate;//保质期
 	int C_amount;//数量
 	char batch;//批次，仅在货架上使用
 } C_goodsInfo;
 typedef struct {//商品信息
-	int C_out;//标价（非实售价）
+	float C_out;//标价（非实售价）
 	char name[C_NAME_LEN];//商品名
 	char code[C_CODE_LEN];//编码
 	int C_shelfDate_Pre;//备选保质期，用于下次录入该商品信息时的自动填充
@@ -30,7 +30,7 @@ typedef struct {
 	float C_in;//进价
 	float C_out;//标价
 }C_goodsReturnPrice;
-C_Goods C_GoodsCreat(int out, char name[], char code[]);//创建C_Goods
+C_Goods C_GoodsCreat(float out, char name[], char code[]);//创建C_Goods
 int C_goodsShelfInfoNum(C_Goods* x);//获取货架商品批次数量
 int C_goodsStockInfoNum(C_Goods* x);//获取库存商品批次数量
 void C_goodsListInit(C_Goods* goods);//库存货架列表初始化
